Export guest list to excel table
I want to directly convert the guest list to excel, because the current csv implementation needs extra excel import to split in the correct columns. Exporting directly to excel format will add extra ease of use for the champions. I would leave the current export button, rename it to (CSV) and add extra new button with name (XLS).

Not fully ready yet, need to figure out how the data can be easily converted to rows, but a first step to check if this works in the API...

with the use of http://stackoverflow.com/questions/17450412/how-to-create-an-excel-file-with-nodejs#17450941